Add REUSE pipeline and missing license
Add REUSE scan to meta-ohos repository as described on
https://git.ostc-eu.org/docs/ci#reuse

Note that this does not scan the sources recursively yet. It's just a starting
point. I suspect passing a recursive scan right now might be difficult as our
manifest includes yocto, openembedded and other third-party remotes and it is
unlikely they are fully REUSE compliant yet.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>