default manifest: point to OSTC fork of meta-zephyr
meta-zephyr needs to be modified for our purposes. Therefore,
a clone is maintained in OSTC git server. The staging branch
is the default and should contain current master + our patches.
The master branch is 1:1 mirror of yoctoproject.org/meta-zephyr
master. Hopefully patches introduced to our staging will be
upstreamed and merged to original meta-zephyr.

Remote `origin` in this manifest pointed to `incubate` group
that does not contain any repositories. Everything related
to yocto is kept in `meta-ohos` group, so fix the remote to
point there.

Signed-off-by: Wojciech Zmuda <wojciech.zmuda@huawei.com>
Change-Id: Ifa4db6ef30af9c0ca8cfbff389694c1b08388276