Switch to a local pipeline, fix linux build
To simplify debugging, switch from the shared pipeline to a local pipeline that
understands the structure of the project. This may be more elaborate once we
get trigers from other repositories (e.g. rev meta-ohos and perform a build by
triggering manifest with some parameters).

In addition re-name the build task to build-linux, as there will be similar
task for Zephyr and FreeRTOS right next to it.

In addition, explain what is going on in the build pipeline

The build is not tested yet.

Fixes: https://git.ostc-eu.org/OSTC/OHOS/manifest/-/issues/6
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>