Generalize CI to cover linux, zephyr and freertos
Most of the build logic around bootstrapping the workspace and setting up
bitbake is identical and was adapted without changes. Three variables govern
how the ".build" job template operates:

- OHOS_BUILD_FLAVOUR picks the "flavour" of OHOS to build: ranging from linux,
  zephyr to freertos. This maps to the flavours of the meta-ohos layer.
- OHOS_BUILD_DIR holds the directory used for the build.
- OHOS_BUILD_GOAL is the bitbake goal to reach.

The only outlier are the artifacts section, as zephyr uses "tmp-newlib" rather
than tmp and after thinking about it, I think this is fine, as we can pick
precise artifacts without having to use more variables and be artificially
generic.

Closes: https://git.ostc-eu.org/OSTC/OHOS/manifest/-/issues/2
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>