Build and execute all zephyr test programs
Following instructions from meta-zephyr README.txt referenced below, build run
and collect results from all the Zephyr test programs. The results are simply
collected and block the build on failure. The results are not adapted to be
compatible with GitLab unit test report format, that is left as a TODO.

Refers-To: http://git.yoctoproject.org/cgit/cgit.cgi/meta-zephyr/tree/README.txt
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>