Introduce "develop.xml" manifest file
Add develop.xml which will incorporate the latest
versions of the repositories under our control:
* OHOS components
* 3rd party layers forks

The reason behind that is to be able to synchronize
with the latest version of the software using "repo"
utility.

It is achieved by removing "remote" and "revision"
from the entries that we want to fall back to the "default"
which is "develop" branch in the "ohos" remote repository.

Signed-off-by: Zbigniew Bodek <zbigniew.bodek@huawei.com>