Extract TMPDIR from bitbake -e
We've noticed that Zephyr do not have any corresponding images. This is
because the set of collected artifacts is now limited to just
tmp/deploy/{images,licenses}, but the "tmp" directory is a variable that
differs per target.

For Zephyr builds it is actually tmp-newlib, but instead of special
casing that, ask Bitbake for the appropriate value.

Related-to: https://git.ostc-eu.org/OSTC/OHOS/manifest/-/issues/18
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>