Define shared .build job, restrict job startup
Move build definitions to build-generic.yaml and further refine them in
the top-level .gitlab-ci.yml. This allows us to spell once, that we want
to restrict CI only to merge requests.

The semantics of merging job definitions is rather simple but forces our
hand to make some repetitions. Dictionary-like values are merged with
older key-value pairs being overridden by clashing keys from the new
definition. Sequence-like values are simply replaced.

The original .build job from .gitlab-ci.yml is reduced to a customized
environment variable and customized build rules, that are somewhat
unique in the manifest repository, as both default.xml and develop.xml
are tested when the relevant files change.

Closes: https://git.ostc-eu.org/OSTC/OHOS/manifest/-/issues/20
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>