Add generic extension mechanism for local.conf
Since changing local.conf seems like the bulk of what various
customizations are about, it makes sense to add a generic way to do that
and avoid having to write more custom code.

Any environment variable starting with OHOS_BB_LOCAL_CONF_ is converted
into a variable assignment in local.conf, with the prefix stripped and
quoting applied, so that all the values are strings.

Nothing is using that, as this has to be coordinated with meta-ohos and
xts-acts repositories.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>