Disable collection of image artifacts
The CI for the manifest repository performs build tests,
there are no boot tests or any other runtime tests that would make use
of the artifacts that are currently collected and stored.

Since each artifact is rather large and they all bounce through the
GitLab instance for storage, it takes a good chunk of time to store
them.

Until there's an actual need for that, stop storing them.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>