Convert public/private cache to OHOS_BB_LOCAL_CONF_
OHOS_TAINTED_BUILD is gone, and is replaced by a set of
OHOS_BB_LOCAL_CONF_ variables and a related helper OHOS_BUILD_CACHE set
to either "public" or "private". All alterations of local.conf are
converted to OHOS_BB_LOCAL_CONF_xxx variables, which further simplifies
the .build and .workspace jobs, to the point where we can consider
merging them in the near future.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>