develop.xml: add docs repository
OpenHarmony documentation is transforming into a model where final
documentation is assembled out of parts located in their respective
repositories, for example meta-ohos or the manifest repository.

This requires the "docs" repository to be represented in the manifest,
so that it has a well-defined, fixed location relative to the other
sources.

Once this process is established it should be replicated in the default
manifest.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>