build-generic.yaml: add support for OHOS_CI_GIT_REPO_PATH
In the trend to move all CI features to the top-level manifest
repository, and only enable and control said features in specific other
repositories, generalize the mechanism used in meta-ohos, to update the
effective revision that is checked out by git-repo.

By setting OHOS_CI_GIT_REPO_PATH, one can update the specific
sub-directory of checked out sources, to $CI_COMMIT_SHA, allowing
testing of incoming changes. This is applicable to CI in repositories
referenced by the manifest repository.

Note that this mechanism is currently limited to branches on the
original remote. This limitation will be addressed separately.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>