build-generic.yaml: add support for OHOS_CI_DEVTOOL_{RECIPE_NAME,LAYER_PATH}
In the trend to move all CI features to the top-level manifest
repository, and only enable and control said features in specific other
repositories, generalize the mechanism used in xts-acts, to update a
git-based recipe stored in a specific layer.

By setting the pair of optional variables, a specific recipe in a
specific layer is updated, with "devtool upgrade", to point to another
git revision. This is applicable to CI in repositories packaged in one
of the layers referenced by the manifest repository.

Note that this mechanism is currently limited to branches on the
original remote. This limitation will be addressed separately.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>