develop.xml: rename to default.xml
The role of separate manifests is now delegated to separate git
branches. This is less confusing to users, less surprising to `repo`,
since `-m develop.xml` is not default and is not "sticky" after `repo
init`, but and equally powerful.

Bulk of the change is related to updates to the CI process. With the
removal separate manifests half of the job definitions can be removed.
For compatibility with the current include statements in meta-ohos and
xts_acts, the develop-targets.yaml file is kept for the moment. I will
follow up with coordinated changes that will allow us to remove it
entirely.

There's a small change to the actual manifest, to arrange "floating"
entries after all the pinned entires, to make the release process easier
to do correctly.

Closes: https://git.ostc-eu.org/OSTC/OHOS/manifest/-/issues/24
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>