build-generic.yaml: merge all pipelines into one file
This step simplifies process of including and correctly configuring the
pipeline in other repositories. It will also unlock refactoring, mainly
aimed at removing the .build job entirely, and moving the precise
functionality to one of the flavour-specific jobs,
.build-{linux,zephyr,freertos} instead.

For backwards compatibility, the two empty files are retained
to allow older pipelines not to fail on missing import file.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>