build-generic.yaml: support layer-level CI from forks
Due to the manifest-based workflow, the GitLab defaults are not
sufficient to support merge requests from forked repositories.

For CI on repositories like meta-ohos, which require replacement of a
the checked out revision of a specific git repository, all that is
needed is to add another remote, here called incoming, and `git fetch`,
that allows the subsequent `git checkout` to work.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>