gitlab-ci.yml: run all the builds when tagged
The patch 7110ae68d41c62a2c4351fb20c1e3cea7ab227c4 "build-generic.yaml: run
build pipeline for incoming tags" intended to have the generic .build
job run when a tag is placed.

Due to the fact that .gitlab-ci.yml re-defines the .build job with
slightly different rules and the fact that rules are defined as a
sequence, GitLab is not merging them in any way, so that tag rule is
lost.

Add the tag rule explicitly at the level of overridden .gitlab-ci.yml.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>