default.xml: Unpin ip-policy project
We initially pinned it because we thought that explicit bumps would not
be needed often. The issue though comes when updating the docs through
the CI when changes are done in the ip-policy repository. If the project
is pinned, docs aggregation will not pick anything new because the
manifest, which is used when creating the docs workspace, will not
change for ip-policy.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>
Change-Id: Ic527548bf172fafeef5cbe486dc6581050e93580