docs: move CI documentation
Documentation for the CI process was started in the docs repository but
we always wanted to have it here. Since we now have a way to do that,
let's make that happen.

To complete the picture offer two new reusable CI jobs - .build-docs and
.aggregate-docs, allowing to express the common scenario of hosting
local documentation and requesting an aggregated build through the docs
repository.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>