build-generic.yaml: Fix workspace setup on fork pushes
When an MR is raised from a fork, GitLab does a merge operation from
source to target. The resulting SHA is used as CI_COMMIT_SHA, which
reside neither in source nor in target - it's a temporary merge in the
scope of the CI validation.

This change fixes the assumption that CI_COMMIT_SHA is in source by
reusing the setup GitLab done in CI_PROJECT_DIR.

Signed-off-by: Andrei Gherzan <andrei@gherzan.com>