.ostc-ci: fix detection of conflicting caching options
The intent was to detect when we both have public cache and accept the
FSL EULA but the shell logic was just broken. Instead of writing:

    if command; then ... fi

I wrote:

    if [ command ]; then ... fi

And nothing in the system noticed, because, separately, none of the
scripts seem to execute with `set -e` in effect. That will be addressed
separately.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>
Change-Id: I6e87261b831ced591466139dbf10316610d4a879