.ostc-ci: fix typo preventing usage of git-repo cache
During one of the OSTC/OHOS rename sprees I must have made the mistake
of renaming {OSTC,OHOS}_MANIFEST_MIRROR variable partially, having the
definition right but the usage wrong. This resulted in CI not really
using the git cache at all, making everything extremely slow if the
remote git repositories were under load.

There's a separate mystery as to why this was not breaking the script on
undefined variable, but I will address that separately.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>
Change-Id: Ie81dd232a782f78cc646fdcee1ca1f31cec5ce11