.ostc-ci: internally include jobs from meta-ohos
The meta-ohos repository contains definitions for supported MACHINE and
FLAVOUR variants but, up until now, the job definitions for building
them were defined in the manifest repository.

Fix this situation by using the recently-added build jobs from the
meta-ohos repository. This is an internal change and it does not affect
any existing pipelines.

Closes: https://git.ostc-eu.org/OSTC/OHOS/manifest/-/issues/32
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>
Change-Id: I5382ea457e273df56d7cc8f8f67e22026e5b2228