Add generic jobs for test and reporting for CI pipeline.
This will add a test-qemu generic job for gitlab which can be
extended with a couple of variables to submit a jobs to LAVA.
LAVA will pick up artifacts from build step and deploy them to qemu
then run the tests.
The report step is manual and will be triggered from the LAVA side
and collect the results from LAVA and into the gitlab.

Signed-off-by: Stevan Radaković <stevan.radakovic@linaro.org>