default.xml: Un-pin meta-zephyr
This layer is very much alive, getting constantly updated with new
upstream and local patches. Un-pin it from specific revision to
make it easier for the new changes to be included in the system
and to avoid desynchronization problem, where another repo
follows meta-zephyr with related changes but manifest is
not updated and the build fails.

Signed-off-by: Wojciech Zmuda <wojciech.zmuda@huawei.com>
Change-Id: I26781208c388cda9a63957f5f7cbccdbc92d9576