Fix scheduled pipelines
Recent work to enable pipelines from forks accidentally broke scheduled
pipelines. This happened because we override OHOS_MANIFEST_URL to point
to the address of the origin remote of a pull request. That variable is
entirely unset in a scheduled pipeline or in several other cases.

This is an attempt at correcting that. The override moves to the
conditional rules section of specific jobs, keeping the defaults intact.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>