default.xml: Switch meta-openembedded remote back to upstream and bump rev
We switched to a local fork of meta-openembedded for a while due to a
fix in nss which was not easy to carry in our integration layer. The
fix is merged upstream and backported to dunfell now so we can switch
the remote again and bump to the latest HEAD of the dunfell branch.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>